+++ trunk/src/common/stepd_api.c
@@ -108,7 +108,7 @@ _handle_stray_socket(const char *socket_name)
 
 	if ((uid = getuid()) != buf.st_uid) {
 		debug3("_handle_stray_socket: socket %s is not owned by uid %d",
-		       socket_name, uid);
+		       socket_name, (int)uid);
 		return;
 	}
